move FilePIcker to Elements directory; fix bugs
