+++ src/Elements/FilePicker.js
@@ -1,11 +1,25 @@
 import React, { Component } from 'react'
 
+// Single value FilePicker. Needs adjustment to support multi-file picking (or new component)
+
+/* Example:
+ * <FilePicker
+ *   id="Logdirectory"
+ *   name="logDirectory"
+ *   defaultValue="Select a Directory"
+ *   buttonValue="Pick a Folder"
+ *   directoriesOnly={true}
+ *   value={this.state.logDirectory}
+ *   onChangeFunction={(value, e) => this.changeLogDirectory(value, e)}
+ * />
+ */
+
 class FilePicker extends Component {
   constructor (props) {
     super(props)
 
     this.state = {
-      logDirectory: props.value || props.defaultValue
+      value: props.value || props.defaultValue
     }
   }
 
@@ -21,13 +35,13 @@ class FilePicker extends Component {
     this.props.onChangeFunction(value, e)
 
     this.setState({
-      logDirectory: value
+      file: value
     })
   }
 
   render () {
     return (
-      <div className="FilePicker">
+      <div className={`FilePicker ${this.props.className || ''}`}>
         <button
           htmlFor={this.props.id}
           onClick={this.clickPickerButton}
@@ -35,14 +49,14 @@ class FilePicker extends Component {
           {this.props.buttonValue}
         </button>
         <div>
-          <p>{this.state.logDirectory}</p>
-          <span className="tooltiptext">{this.state.logDirectory}</span>
+          <p>{this.state.value}</p>
+          <span className="tooltiptext">{this.state.value}</span>
         </div>
         <input
           type="file"
           name={this.props.name}
           id={this.props.id}
-          webkitdirectory="true"
+          webkitdirectory={this.props.directoriesOnly ? "true" : null}
           ref={input => this.pickerButton = input}
           onChange={this.changePicker}
         />
